#191 Group by Docker Compose
Each container have a Project which may represent a Docker Compose project.
The name of project is simply the name of folder from which `docker-compose up` executed.
If container was started not from Compose then it anyway will be assigned to the noneProject with empty name.

Each project will be rendered as a separate CompactRow widget.
Currently to determine that this line is a project instead of status will be shown a down arrow ▼. TBD: use ident instead.
All containers will be sorted by the project first and only then by name.
On rendering in grid.go we know that containers from the same project are always together.
Rendered result looks like:

 ▼
  ◉ container1
  ◉ container2
 ▼ proj1
  ◉ proj1_db
  ◉ proj1_serv
 ▼ proj2
  ◉ proj2_db
  ◉ proj2_serv