Update README.md: fix Debian/Ubuntu installation instruction
As explained in issue https://github.com/bcicen/ctop/issues/179, jammy (Ubuntu 22.04.2 LTS) is unknown from the `azlux.fr` repo and it will prevent people to install `ctop` on `Ubuntu 22.04` and later. Also we can see in the script http://packages.azlux.fr/scripts/docker-ctop.txt that the only version is the `stabe` version. So we should use `stable` in the installation commands.